refactor(drift_sync): extract formatters.py (6/8)
Move 5 format_* functions + _group_by_severity + _count_by_severity +
_SEVERITY_ORDER into formatters.py. Pure rendering, no I/O.

Unused `json as _json` import cleaned from __init__.py.

Refs #47 (Theme A item 4).
